code refactoring (#129)
* :hammer: code refactoring

* :hammer: format code

* :hammer: pump up the version

* :green_heart: fix the problem with the import. mysteriously, moban.extensions works but relocating it to moban.jinja2.extension wont do

* :hammer: code refactoring

* :sparkles: code refactoring

* :hammer: code refactoring on jinja2 engine

* :hammer: code refactoring

* :hammer: load handle_bar by default for now

* :green_heart: include pybars3 as temp requirements to pass unit tests

* :bug: fix a typo

* :hammer: allow base engine to have constructor that has more shared code

* :tructor: rename engine_factory

* :hammer: code refactor jinja2 engine

* :fire: remove duplicated find template file

* :fire: remove useless function

* :fire: remove logging
