+++ .moban.cd/moban.yml
@@ -3,8 +3,8 @@ organisation: moremoban
 author: C. W.
 contact: wangc_2011@hotmail.com
 license: MIT
-version: 0.3.3
-current_version: 0.3.3
+version: 0.3.4
+current_version: 0.3.4
 release: 0.3.3
 branch: master
 command_line_interface: "moban"
@@ -19,5 +19,6 @@ dependencies:
   - jinja2>=2.7.1
   - lml==0.0.4
   - crayons
+  - pybars3
 description: Yet another jinja2 cli command for static text generation
 scm_host: github.com
