+++ moban/_version.py
@@ -1,2 +1,2 @@
-__version__ = "0.3.3"
+__version__ = "0.3.4"
 __author__ = "C. W."
