+++ moban/extensions.py
@@ -1,21 +1,6 @@
-from lml.plugin import PluginInfo, PluginManager
+from lml.plugin import PluginInfo
 
-import moban.constants as constants
-
-
-class PluginMixin:
-    def get_all(self):
-        for name in self.registry.keys():
-            # only the first matching one is returned
-            the_filter = self.load_me_now(name)
-            yield (name, the_filter)
-
-
-class JinjaFilterManager(PluginManager, PluginMixin):
-    def __init__(self):
-        super(JinjaFilterManager, self).__init__(
-            constants.JINJA_FILTER_EXTENSION
-        )
+from moban import constants
 
 
 class JinjaFilter(PluginInfo):
@@ -24,11 +9,6 @@ class JinjaFilter(PluginInfo):
 
     def tags(self):
         yield self.cls.__name__
-
-
-class JinjaTestManager(PluginManager, PluginMixin):
-    def __init__(self):
-        super(JinjaTestManager, self).__init__(constants.JINJA_TEST_EXTENSION)
 
 
 class JinjaTest(PluginInfo):
@@ -48,22 +28,6 @@ def jinja_tests(**keywords):
         JinjaTest(key)(value)
 
 
-class JinjaGlobalsManager(PluginManager, PluginMixin):
-    def __init__(self):
-        super(JinjaGlobalsManager, self).__init__(
-            constants.JINJA_GLOBALS_EXTENSION
-        )
-
-
 def jinja_global(identifier, dict_obj):
     plugin = PluginInfo(constants.JINJA_GLOBALS_EXTENSION, tags=[identifier])
     plugin(dict_obj)
-
-
-class LibraryManager(PluginManager):
-    def __init__(self):
-        super(LibraryManager, self).__init__(constants.LIBRARY_EXTENSION)
-
-    def resource_path_of(self, library_name):
-        library = self.get_a_plugin(library_name)
-        return library.resources_path
