+++ moban/jinja2/engine.py
@@ -6,29 +6,18 @@ from lml.plugin import PluginInfo
 import moban.utils as utils
 import moban.reporter as reporter
 import moban.constants as constants
+from moban import plugins
 from moban.utils import get_template_path
 from moban.hashstore import HASH_STORE
-from moban.base_engine import BaseEngine
-from moban.engine_factory import (
-    Context,
-    verify_the_existence_of_directories,
-    Strategy
-)
-from moban import plugins
 
 
 @PluginInfo(
     constants.TEMPLATE_ENGINE_EXTENSION, tags=["jinja2", "jinja", "jj2", "j2"]
 )
-class Engine(BaseEngine):
+class Engine(plugins.BaseEngine):
     def __init__(self, template_dirs, context_dirs):
-        BaseEngine.__init__(self)
-        plugins.refresh_plugins()
-        template_dirs = list(
-            plugins.expand_template_directories(template_dirs))
-        verify_the_existence_of_directories(template_dirs)
-        context_dirs = plugins.expand_template_directory(context_dirs)
-        template_loader = FileSystemLoader(template_dirs)
+        super(Engine, self).__init__(template_dirs, context_dirs)
+        template_loader = FileSystemLoader(self.template_dirs)
         self.jj2_environment = Environment(
             loader=template_loader,
             keep_trailing_newline=True,
@@ -44,35 +33,20 @@ class Engine(BaseEngine):
         for global_name, dict_obj in plugins.GLOBALS.get_all():
             self.jj2_environment.globals[global_name] = dict_obj
 
-        self.context = Context(context_dirs)
-        self.template_dirs = template_dirs
-
     def render_to_file(self, template_file, data_file, output_file):
-        template = self.jj2_environment.get_template(template_file)
         data = self.context.get_data(data_file)
-        template.globals['__target__'] = output_file
-        template.globals['__template__'] = template.name
+        template_file, template = self._get_jinja2_template(template_file)
+        self._apply_template(template_file, template, data, output_file)
         reporter.report_templating(template_file, output_file)
-
-        rendered_content = template.render(**data)
-        utils.write_file_out(output_file, rendered_content)
-        self._file_permissions_copy(template_file, output_file)
-
-    def render_to_files(self, array_of_param_tuple):
-        sta = Strategy(array_of_param_tuple)
-        sta.process()
-        choice = sta.what_to_do()
-        if choice == Strategy.DATA_FIRST:
-            self._render_with_finding_data_first(sta.data_file_index)
-        else:
-            self._render_with_finding_template_first(sta.template_file_index)
 
     def _render_with_finding_template_first(self, template_file_index):
         for (template_file, data_output_pairs) in template_file_index.items():
-            template = self.jj2_environment.get_template(template_file)
+            template_file, template = self._get_jinja2_template(template_file)
             for (data_file, output) in data_output_pairs:
                 data = self.context.get_data(data_file)
-                flag = self._apply_template(template, data, output)
+                flag = self._apply_template(
+                    template_file, template, data, output
+                )
                 if flag:
                     reporter.report_templating(template_file, output)
                     self.templated_count += 1
@@ -82,8 +56,12 @@ class Engine(BaseEngine):
         for (data_file, template_output_pairs) in data_file_index.items():
             data = self.context.get_data(data_file)
             for (template_file, output) in template_output_pairs:
-                template = self.jj2_environment.get_template(template_file)
-                flag = self._apply_template(template, data, output)
+                template_file, template = self._get_jinja2_template(
+                    template_file
+                )
+                flag = self._apply_template(
+                    template_file, template, data, output
+                )
                 if flag:
                     reporter.report_templating(template_file, output)
                     self.templated_count += 1
@@ -97,19 +75,25 @@ class Engine(BaseEngine):
                 break
         utils.file_permissions_copy(true_template_file, output_file)
 
-    def _apply_template(self, template, data, output):
-        template.globals['__target__'] = output
-        template.globals['__template__'] = template.name
-        temp_file_path = get_template_path(self.template_dirs, template)
+    def _apply_template(self, template_file, template, data, output):
+        template.globals["__target__"] = output
+        template.globals["__template__"] = template.name
         rendered_content = template.render(**data)
         rendered_content = utils.strip_off_trailing_new_lines(rendered_content)
         rendered_content = rendered_content.encode("utf-8")
         flag = HASH_STORE.is_file_changed(
-            output, rendered_content, temp_file_path
+            output, rendered_content, template_file
         )
         if flag:
             utils.write_file_out(
                 output, rendered_content, strip=False, encode=False
             )
-            utils.file_permissions_copy(temp_file_path, output)
+            utils.file_permissions_copy(template_file, output)
         return flag
+
+    def _get_jinja2_template(self, template_file):
+        actual_template_file = get_template_path(
+            self.template_dirs, template_file
+        )
+        template = self.jj2_environment.get_template(template_file)
+        return actual_template_file, template
