+++ moban/mobanfile.py
@@ -7,6 +7,7 @@ from lml.utils import do_import
 
 import moban.reporter as reporter
 import moban.constants as constants
+from moban import plugins
 from moban.utils import (
     merge,
     git_clone,
@@ -15,7 +16,6 @@ from moban.utils import (
     expand_directories,
 )
 from moban.copier import Copier
-from moban import plugins
 
 try:
     from urllib.parse import urlparse
