+++ setup.py
@@ -5,13 +5,15 @@ import os
 import sys
 import codecs
 from shutil import rmtree
-from setuptools import setup, find_packages, Command
+
+from setuptools import Command, setup, find_packages
+
 PY2 = sys.version_info[0] == 2
 PY26 = PY2 and sys.version_info[1] < 7
 
 NAME = 'moban'
 AUTHOR = 'C. W.'
-VERSION = '0.3.3'
+VERSION = '0.3.4'
 EMAIL = 'wangc_2011@hotmail.com'
 LICENSE = 'MIT'
 ENTRY_POINTS = {
@@ -26,9 +28,9 @@ URL = 'https://github.com/moremoban/moban'
 DOWNLOAD_URL = '%s/archive/0.3.3.tar.gz' % URL
 FILES = ['README.rst', 'CONTRIBUTORS.rst', 'CHANGELOG.rst']
 KEYWORDS = [
+    'python',
     'jinja2',
     'moban',
-    'python'
 ]
 
 CLASSIFIERS = [
@@ -48,6 +50,7 @@ INSTALL_REQUIRES = [
     'jinja2>=2.7.1',
     'lml==0.0.4',
     'crayons',
+    'pybars3',
 ]
 SETUP_COMMANDS = {}
