+++ tests/integration_tests/test_command_line_options.py
@@ -3,7 +3,6 @@ import sys
 from shutil import copyfile
 
 from mock import patch
-from moban.main import main
 from nose.tools import eq_, raises, assert_raises
 
 
@@ -13,15 +12,11 @@ class TestCustomOptions:
         with open(self.config_file, "w") as f:
             f.write("hello: world")
         self.patcher1 = patch(
-            "moban.engine.verify_the_existence_of_directories"
-        )
-        self.patcher2 = patch(
-            "moban.engine_factory.verify_the_existence_of_directories"
+            "moban.plugins.verify_the_existence_of_directories"
         )
         self.patcher1.start()
-        self.patcher2.start()
 
-    @patch("moban.engine.Engine.render_to_file")
+    @patch("moban.jinja2.engine.Engine.render_to_file")
     def test_custom_options(self, fake_template_doer):
         test_args = [
             "moban",
@@ -35,15 +30,19 @@ class TestCustomOptions:
             "a.jj2",
         ]
         with patch.object(sys, "argv", test_args):
+            from moban.main import main
+
             main()
             fake_template_doer.assert_called_with(
                 "a.jj2", "config.yaml", "moban.output"
             )
 
-    @patch("moban.engine.Engine.render_to_file")
+    @patch("moban.jinja2.engine.Engine.render_to_file")
     def test_minimal_options(self, fake_template_doer):
         test_args = ["moban", "-c", self.config_file, "-t", "a.jj2"]
         with patch.object(sys, "argv", test_args):
+            from moban.main import main
+
             main()
             fake_template_doer.assert_called_with(
                 "a.jj2", "config.yaml", "moban.output"
@@ -53,11 +52,12 @@ class TestCustomOptions:
     def test_missing_template(self):
         test_args = ["moban", "-c", self.config_file]
         with patch.object(sys, "argv", test_args):
+            from moban.main import main
+
             main()
 
     def tearDown(self):
         self.patcher1.stop()
-        self.patcher2.stop()
         os.unlink(self.config_file)
 
 
@@ -67,18 +67,16 @@ class TestOptions:
         with open(self.config_file, "w") as f:
             f.write("hello: world")
         self.patcher1 = patch(
-            "moban.engine.verify_the_existence_of_directories"
-        )
-        self.patcher2 = patch(
-            "moban.engine_factory.verify_the_existence_of_directories"
+            "moban.plugins.verify_the_existence_of_directories"
         )
         self.patcher1.start()
-        self.patcher2.start()
 
-    @patch("moban.engine.Engine.render_to_file")
+    @patch("moban.jinja2.engine.Engine.render_to_file")
     def test_default_options(self, fake_template_doer):
         test_args = ["moban", "-t", "a.jj2"]
         with patch.object(sys, "argv", test_args):
+            from moban.main import main
+
             main()
             fake_template_doer.assert_called_with(
                 "a.jj2", "data.yml", "moban.output"
@@ -88,11 +86,12 @@ class TestOptions:
     def test_no_argments(self):
         test_args = ["moban"]
         with patch.object(sys, "argv", test_args):
+            from moban.main import main
+
             main()
 
     def tearDown(self):
         self.patcher1.stop()
-        self.patcher2.stop()
         os.unlink(self.config_file)
 
 
@@ -100,6 +99,8 @@ class TestOptions:
 def test_missing_configuration():
     test_args = ["moban", "-t", "a.jj2"]
     with patch.object(sys, "argv", test_args):
+        from moban.main import main
+
         main()
 
 
@@ -114,18 +115,16 @@ class TestNoOptions:
         with open(self.data_file, "w") as f:
             f.write("hello: world")
         self.patcher1 = patch(
-            "moban.engine.verify_the_existence_of_directories"
-        )
-        self.patcher2 = patch(
-            "moban.engine_factory.verify_the_existence_of_directories"
+            "moban.plugins.verify_the_existence_of_directories"
         )
         self.patcher1.start()
-        self.patcher2.start()
 
-    @patch("moban.engine.Engine.render_to_files")
+    @patch("moban.jinja2.engine.Engine.render_to_files")
     def test_single_command(self, fake_template_doer):
         test_args = ["moban"]
         with patch.object(sys, "argv", test_args):
+            from moban.main import main
+
             main()
             call_args = list(fake_template_doer.call_args[0][0])
             eq_(
@@ -136,10 +135,12 @@ class TestNoOptions:
                 ],
             )
 
-    @patch("moban.engine.Engine.render_to_files")
+    @patch("moban.jinja2.engine.Engine.render_to_files")
     def test_single_command_with_a_few_options(self, fake_template_doer):
         test_args = ["moban", "-t", "abc.jj2", "-o", "xyz.output"]
         with patch.object(sys, "argv", test_args):
+            from moban.main import main
+
             main()
             call_args = list(fake_template_doer.call_args[0][0])
             eq_(
@@ -151,7 +152,7 @@ class TestNoOptions:
                 ],
             )
 
-    @patch("moban.engine.Engine.render_to_files")
+    @patch("moban.jinja2.engine.Engine.render_to_files")
     def test_single_command_with_options(self, fake_template_doer):
         test_args = [
             "moban",
@@ -163,6 +164,8 @@ class TestNoOptions:
             "xyz.output",
         ]
         with patch.object(sys, "argv", test_args):
+            from moban.main import main
+
             main()
             call_args = list(fake_template_doer.call_args[0][0])
             eq_(
@@ -175,17 +178,18 @@ class TestNoOptions:
             )
 
     @raises(Exception)
-    @patch("moban.engine.Engine.render_to_files")
+    @patch("moban.jinja2.engine.Engine.render_to_files")
     def test_single_command_without_output_option(self, fake_template_doer):
         test_args = ["moban", "-t", "abc.jj2"]
         with patch.object(sys, "argv", test_args):
+            from moban.main import main
+
             main()
 
     def tearDown(self):
         os.unlink(self.config_file)
         os.unlink(self.data_file)
         self.patcher1.stop()
-        self.patcher2.stop()
 
 
 class TestNoOptions2:
@@ -199,18 +203,16 @@ class TestNoOptions2:
         with open(self.data_file, "w") as f:
             f.write("hello: world")
         self.patcher1 = patch(
-            "moban.engine.verify_the_existence_of_directories"
-        )
-        self.patcher2 = patch(
-            "moban.engine_factory.verify_the_existence_of_directories"
+            "moban.plugins.verify_the_existence_of_directories"
         )
         self.patcher1.start()
-        self.patcher2.start()
 
-    @patch("moban.engine.Engine.render_to_files")
+    @patch("moban.jinja2.engine.Engine.render_to_files")
     def test_single_command(self, fake_template_doer):
         test_args = ["moban"]
         with patch.object(sys, "argv", test_args):
+            from moban.main import main
+
             main()
             call_args = list(fake_template_doer.call_args[0][0])
             eq_(
@@ -223,7 +225,6 @@ class TestNoOptions2:
 
     def tearDown(self):
         self.patcher1.stop()
-        self.patcher2.stop()
         os.unlink(self.config_file)
         os.unlink(self.data_file)
 
@@ -238,18 +239,16 @@ class TestCustomMobanFile:
         with open(self.data_file, "w") as f:
             f.write("hello: world")
         self.patcher1 = patch(
-            "moban.engine.verify_the_existence_of_directories"
-        )
-        self.patcher2 = patch(
-            "moban.engine_factory.verify_the_existence_of_directories"
+            "moban.plugins.verify_the_existence_of_directories"
         )
         self.patcher1.start()
-        self.patcher2.start()
 
-    @patch("moban.engine.Engine.render_to_files")
+    @patch("moban.jinja2.engine.Engine.render_to_files")
     def test_single_command(self, fake_template_doer):
         test_args = ["moban", "-m", self.config_file]
         with patch.object(sys, "argv", test_args):
+            from moban.main import main
+
             main()
             call_args = list(fake_template_doer.call_args[0][0])
             eq_(
@@ -262,17 +261,18 @@ class TestCustomMobanFile:
 
     def tearDown(self):
         self.patcher1.stop()
-        self.patcher2.stop()
         os.unlink(self.config_file)
         os.unlink(self.data_file)
 
 
-@patch("moban.engine.verify_the_existence_of_directories")
+@patch("moban.plugins.verify_the_existence_of_directories")
 def test_duplicated_targets_in_moban_file(fake_verify):
     config_file = "duplicated.moban.yml"
     copyfile(os.path.join("tests", "fixtures", config_file), ".moban.yml")
     test_args = ["moban"]
     with patch.object(sys, "argv", test_args):
+        from moban.main import main
+
         assert_raises(SystemExit, main)
     os.unlink(".moban.yml")
 
@@ -282,30 +282,36 @@ class TestInvalidMobanFile:
         self.config_file = ".moban.yml"
 
     @raises(SystemExit)
-    @patch("moban.engine.Engine.render_to_files")
+    @patch("moban.jinja2.engine.Engine.render_to_files")
     def test_no_configuration(self, fake_template_doer):
         with open(self.config_file, "w") as f:
             f.write("")
         test_args = ["moban"]
         with patch.object(sys, "argv", test_args):
+            from moban.main import main
+
             main()
 
     @raises(SystemExit)
-    @patch("moban.engine.Engine.render_to_files")
+    @patch("moban.jinja2.engine.Engine.render_to_files")
     def test_no_configuration_2(self, fake_template_doer):
         with open(self.config_file, "w") as f:
             f.write("not: related")
         test_args = ["moban"]
         with patch.object(sys, "argv", test_args):
+            from moban.main import main
+
             main()
 
     @raises(SystemExit)
-    @patch("moban.engine.Engine.render_to_files")
+    @patch("moban.jinja2.engine.Engine.render_to_files")
     def test_no_targets(self, fake_template_doer):
         with open(self.config_file, "w") as f:
             f.write("configuration: test")
         test_args = ["moban"]
         with patch.object(sys, "argv", test_args):
+            from moban.main import main
+
             main()
 
     def tearDown(self):
@@ -322,33 +328,30 @@ class TestComplexOptions:
         with open(self.data_file, "w") as f:
             f.write("hello: world")
         self.patcher1 = patch(
-            "moban.engine.verify_the_existence_of_directories"
-        )
-        self.patcher2 = patch(
-            "moban.engine_factory.verify_the_existence_of_directories"
+            "moban.plugins.verify_the_existence_of_directories"
         )
         self.patcher1.start()
-        self.patcher2.start()
 
-    @patch("moban.engine.Engine.render_to_files")
-    def test_single_command(self, fake_template_doer):
+    def test_single_command(self):
         test_args = ["moban"]
         with patch.object(sys, "argv", test_args):
-            main()
-            call_args = list(fake_template_doer.call_args[0][0])
-            eq_(
-                call_args,
-                [
-                    ("README.rst.jj2", "custom-data.yaml", "README.rst"),
-                    ("setup.py.jj2", "data.yml", "setup.py"),
-                ],
-            )
+            from moban.main import main
+
+            with patch("moban.jinja2.engine.Engine.render_to_files") as fake:
+                main()
+                call_args = list(fake.call_args[0][0])
+                eq_(
+                    call_args,
+                    [
+                        ("README.rst.jj2", "custom-data.yaml", "README.rst"),
+                        ("setup.py.jj2", "data.yml", "setup.py"),
+                    ],
+                )
 
     def tearDown(self):
         os.unlink(self.config_file)
         os.unlink(self.data_file)
         self.patcher1.stop()
-        self.patcher2.stop()
 
 
 class TestTemplateTypeOption:
@@ -357,10 +360,12 @@ class TestTemplateTypeOption:
         with open(self.config_file, "w") as f:
             f.write("hello: world")
 
-    @patch("moban.engine.Engine.render_to_file")
+    @patch("moban.jinja2.engine.Engine.render_to_file")
     def test_mako_option(self, fake_template_doer):
         test_args = ["moban", "-t", "a.mako"]
         with patch.object(sys, "argv", test_args):
+            from moban.main import main
+
             main()
             fake_template_doer.assert_called_with(
                 "a.mako", "data.yml", "moban.output"
